warn fix

+++ services/distributeddataservice/service/data_share/strategies/general/load_config_data_info_strategy.cpp
@@ -82,7 +82,7 @@ bool LoadConfigSingleDataInfoStrategy::operator()(std::shared_ptr<Context> conte
         }
     }
     context->calledSourceDir = metaData.dataDir;
-    context->version = metaData.version;
+    context->version = static_cast<int>(metaData.version);
     return true;
 }
 } // namespace OHOS::DataShare
